refactor: utils methods for eth tests
